refactor: restructured folders
