+++ src/components/carousel/Carousel.css
