+++ src/components/header/Header.js
@@ -51,4 +51,4 @@ const Header = props => {
     </header>
   );
 };
-export default Header;
+export  {Header};
