+++ src/components/header/SearchInput.css
