+++ src/components/index.js
@@ -1,0 +1,9 @@
+export { AuthModal } from './authorization/AuthModal';
+export { Banner } from './banner/Banner';
+export { Carousel } from './carousel/Carousel';
+export { Categories } from './categories/Categories';
+export { Footer } from './footer/Footer';
+export { Header } from './header/Header';
+export { ScrollTopButton } from './scroll-top-button/ScrollTopButton';
+export { Testimonies } from './testimonies/Testimonies';
+export { Modal } from './ui/Modal';
