+++ src/components/scroll-top-button/ScrollTopButton.js
@@ -13,4 +13,4 @@ const ScrollTopButton = () => {
   );
 };
 
-export default ScrollTopButton;
+export { ScrollTopButton };
