+++ src/data/testimonies-data.json
@@ -1,0 +1,29 @@
+[{
+        "id": "testimony-01",
+        "content": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book",
+        "avatar": "https://picsum.photos/536/354",
+        "testimonyName": "Rohan Kapoor",
+        "designation": "CEO, Omega Auto"
+    },
+    {
+        "id": "testimony-02",
+        "content": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book",
+        "avatar": "https://picsum.photos/536/354",
+        "testimonyName": "Mohan Kapoor",
+        "designation": "COO, Balaji Auto"
+    },
+    {
+        "id": "testimony-03",
+        "content": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book",
+        "avatar": "https://picsum.photos/536/354",
+        "testimonyName": "Sohan Kapoor",
+        "designation": "MD, Hero Automobiles"
+    },
+    {
+        "id": "testimony-04",
+        "content": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book",
+        "avatar": "https://picsum.photos/536/354",
+        "testimonyName": "Lakhan Kapoor",
+        "designation": "CEO, Torque Machines"
+    }
+]
